requirements.txt
Django>=3.2.6
celery==5.1.2
redis==3.5.3
psycopg2==2.8.5
djangorestframework==3.11.0
